Add generic Resource startup
