+++ Graphir.API/Extensions/Startup.Services.cs
@@ -44,26 +44,26 @@ public static class StartupServices
             .AddFhirTypes()
             .AddPatient()
             .AddPractitioner()
-            .AddOrganization()
+            .AddResourceType<Organization, OrganizationType>()
             .AddAppointment()
-            .AddMedication()
-            .AddLocation()
-            .AddDevice()
+            .AddResourceType<Medication, MedicationType>()
+            .AddResourceType<Location, LocationType>()
+            .AddResourceType<Device, DeviceType>()
             .AddSchedule()
-            .AddHealthcareService()
-            .AddSlots()
-            .AddMedicationRequest()
-            .AddMedicationAdministration()
-            .AddProvenance()
-            .AddCoverage()
-            .AddCondition()
-            .AddGroup()
-            .AddEndpoint()
-            .AddEncounter()
-            .AddPractitionerRole()
+            .AddResourceType<HealthcareService, HealthcareServiceType>()
+            .AddResourceType<Slot, SlotType>()
+            .AddResourceType<MedicationRequest, MedicationRequestType>()
+            .AddResourceType<MedicationAdministration, MedicationAdministrationType>()
+            .AddResourceType<Provenance, ProvenanceType>()
+            .AddResourceType<Coverage, CoverageType>()
+            .AddResourceType<Condition, ConditionType>()
+            .AddResourceType<Group, GroupType>()
+            .AddResourceType<Endpoint, EndpointType>()
+            .AddResourceType<Encounter, EncounterType>()
+            .AddResourceType<PractitionerRole, PractitionerRoleType>()
             .AddRelatedPerson()
-            .AddDetectedIssue()
-            .AddServiceRequest()
+            .AddResourceType<DetectedIssue, DetectedIssueType>()
+            .AddResourceType<ServiceRequest, ServiceRequestType>()
             .AddResourceType<Specimen, SpecimenType>()
             .AddResourceType<ClinicalImpression, ClinicalImpressionType>()
             .AddResourceType<Substance, SubstanceType>()
@@ -73,13 +73,19 @@ public static class StartupServices
             .AddResourceType<CarePlan, CarePlanType>()
             .AddResourceType<DeviceRequest, DeviceRequestType>()
             .AddResourceType<MedicationDispense, MedicationDispenseType>()
+            .AddResourceType<MedicationStatement, MedicationStatementType>()
             .AddResourceType<Procedure, ProcedureType>()
             .AddResourceType<Immunization, ImmunizationType>()
+            .AddResourceType<ImmunizationRecommendation, ImmunizationRecommendationType>()
             .AddResourceType<ImagingStudy, ImagingStudyType>()
             .AddResourceType<DeviceMetric, DeviceMetricType>()
+            .AddResourceType<Questionnaire, QuestionnaireType>()
             .AddResourceType<QuestionnaireResponse, QuestionnaireResponseType>()
             .AddResourceType<MolecularSequence, MolecularSequenceType>()
             .AddResourceType<DocumentReference, DocumentReferenceType>()
+            .AddResourceType<AllergyIntolerance, AllergyIntoleranceType>()
+            .AddResourceType<Media, MediaType>()
+            .AddResourceType<NutritionOrder, NutritionOrderType>()
             ;
     }
 }
