+++ Graphir.API/Schema/AppointmentType.cs
@@ -82,21 +82,3 @@ public class AppointmentParticipantActorReferenceType : UnionType
         descriptor.Type<LocationType>();
     }
 }
-
-public class AppointmentQuery : ObjectTypeExtension<Query>
-{
-    protected override void Configure(IObjectTypeDescriptor<Query> descriptor)
-    {
-        descriptor.Field("Appointment")
-                .Type<AppointmentType>()
-                .Argument("id", a => a.Type<NonNullType<StringType>>())
-                .ResolveWith<ResourceResolvers<Appointment>>(r => r.GetResource(default!, default!));
-
-        descriptor.Field("AppointmentList")
-            .Type<ListType<AppointmentType>>()
-            .Argument("patient", a => a.Type<StringType>())
-            .Argument("practitioner", a => a.Type<StringType>())
-            .Argument("date", a => a.Type<DateType>())
-            .ResolveWith<AppointmentResolvers>(r => r.GetAppointmentListResources(default!, default, default, default));
-    }
-}
